+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Using string instead of using an object when computing the hash for the url manager key
 
 ## [0.1.0] - 2026-02-25
 - Added "deleteKey" and "deleteAllKeys" methods to the URL Manager